Integrate the code needed to test the missing AOK…
…endpoint.

This commit pulls in test code from UAA to test the /oauth/clients/tx/modify
endpoint which is what we need for BZ#105536.

The code is not from UAA upstream master branch, because too much has changed
and that would be a large task. This code is from the 1.6.0 release tag of
UAA, which is the first release tag after the endpoint was added.

This test results in a 404 from AOK, which is exactly what we want. Next step
is to implement the endpoint in AOK.

After that we can run the test for BZ#105140:

    stackato add-service-broker --username containers --password secret --url http://… cf-containers-broker

I've worked out an exact reproduction of that, so should be easy to verify
BZ#105140 when the endpoint is implemented.